pyproject.toml
[project]
name = "sam3-image-service"
version = "0.1.0"
description = "SAM3 image recognition service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116,<1",
    "numpy>=2.3,<3",
    "pillow>=11,<12",
    "pydantic-settings>=2.10,<3",
    "python-multipart>=0.0.20,<0.1",
    "uvicorn[standard]>=0.35,<1",
]

[project.optional-dependencies]
runtime = [
    "psutil>=7,<8",
    "sam3==0.1.3",
    "torch==2.7.0",
    "torchvision>=0.22,<0.23",
]

[project.scripts]
sam3-image-service = "sam3_image_service:main"

[dependency-groups]
dev = [
    "httpx>=0.28,<0.29",
    "pytest>=8.4,<9",
]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
